Extract `GoogleSignInSafariScreen` in dedicated file (and use English!)
I've been working on this form Italy and set my text to what the Italian
Safari showed while developing... And obviously forgot to update to use
English once done.
